Rename Instance to CVRP; rename solution to parse_solution; new parse_instance file
